Stop testing error paths against a third-party host
PluginManagerTest and PluginManagerDialogTest pointed the repository address
and two download URLs at http://httpstat.us/500, a public service, to exercise
the failure paths. That made the suite depend on someone else's uptime, fail
outright when offline, and spend its time in connect timeouts and retry waits:
PluginManagerTest alone took 221 of the 230 seconds of a local run.

testRetryDownload already had the answer - a loopback HttpServer answering 500 -
so lift that into FailingHttpStub and point all four sites at it. The class drops
to 22 s, nearly all of which is the retry intervals the tests deliberately assert.
